+++ package.json
@@ -5,6 +5,7 @@
     "@types/chrome": "^0.0.123",
     "@types/copy-webpack-plugin": "^6.0.0",
     "@types/webpack": "^4.41.22",
+    "@types/webpack-merge": "^4.1.5",
     "copy-webpack-plugin": "^6.1.1",
     "grunt": "^1.3.0",
     "grunt-shell": "^3.0.1",
@@ -12,7 +13,8 @@
     "ts-node": "^9.0.0",
     "typescript": "^4.0.3",
     "webpack": "^4.44.2",
-    "webpack-cli": "^3.3.12"
+    "webpack-cli": "^3.3.12",
+    "webpack-merge": "^5.1.4"
   },
   "dependencies": {
     "rxjs": "^6.6.3"
